feat: create functionality for menu categories
