+++ main/index.html
@@ -14,6 +14,7 @@
     <link rel="stylesheet" href="./style.css">
     <link rel="stylesheet" href="../index.css">
     <script src="./script.js" defer></script>
+    <script src="./scripts/burger.js" defer></script>
   </head>
   <body>
     <div id="mySidenav" class="sidenav">
