+++ menu/index.html
@@ -14,7 +14,9 @@
     >
     <link rel="stylesheet" href="./style.css" >
     <link rel="stylesheet" href="../index.css" >
-    <script type="module" src="./script.js" defer></script>
+    <script src="./scripts/menu.js" type="module" defer></script>
+    <script src="./scripts/burger.js" defer></script>
+    <script src="./script.js" defer></script>
   </head>
   <body>
     <div id="mySidenav" class="sidenav">
