+++ menu/style.css
@@ -178,7 +178,7 @@ input[type='radio']:checked + label .tabs-wrapper {
     max-width: 90%;
   }
 
-  .product-list .product-wrapper:nth-last-child(-n + 4) {
+  .hidden-class:nth-last-child(-n + 4) {
     display: none;
   }
